requirements.txt
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==9.1.0
django-braces==1.12.0
djangorestframework==3.7.7
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
python-decouple==3.1
pytz==2017.3
whitenoise==3.3.1
